+++ R/add_author.R
@@ -1,9 +1,9 @@
 #' Add an author
 #'
 #' @param study A study list object with class scivrs_study
-#' @param orcid the author's unique ORCiD (see https://orcid.org/)
 #' @param surname a character string with the author's last name(s)
 #' @param given a character string with the author's given name(s)
+#' @param orcid the author's unique ORCiD (see https://orcid.org/)
 #' @param roles a vector of roles from the CRediT taxonomy (see https://casrai.org/credit/); use credit_roles() to view the full list
 #' @param ... further info to add to author object
 #'
@@ -12,27 +12,28 @@
 #'
 #' @examples
 #' s <- study() %>% add_author(
-#'   orcid = "0000-0002-7523-5539",
 #'   surname = "DeBruine", given = "Lisa M.",
+#'   orcid = "0000-0002-7523-5539",
 #'   roles = c("Conceptualization", "Methodology")
 #' )
 #' study_to_json(s)
 #'
-add_author <- function(study, orcid, surname, given = "",
+add_author <- function(study, surname, given = "",
+                       orcid = NULL,
                        roles = c(), ...) {
 
   idx <- get_idx(study, section = "authors")
 
-  study$authors[[idx]] <- author(orcid, surname, given, roles, ...)
+  study$authors[[idx]] <- author(surname, given, orcid, roles, ...)
 
   invisible(study)
 }
 
 #' Author in CRediT Format
 #'
-#' @param orcid the author's unique ORCiD (see https://orcid.org/)
 #' @param surname a character string with the author's last name(s)
 #' @param given a character string with the author's given name(s)
+#' @param orcid the author's unique ORCiD (see https://orcid.org/)
 #' @param roles a vector of roles from the CRediT taxonomy (see https://casrai.org/credit/); use credit_roles() to view the full list
 #' @param ... further info to add to author object
 #'
@@ -41,11 +42,11 @@ add_author <- function(study, orcid, surname, given = "",
 #'
 #' @examples
 #' author(
-#'   orcid = "0000-0002-7523-5539",
 #'   surname = "DeBruine", given = "Lisa M.",
+#'   orcid = "0000-0002-7523-5539",
 #'   roles = c("Conceptualization", "Methodology")
 #' )
-author <- function(orcid, surname, given = "", roles = c(), ...) {
+author <- function(surname, given = "", orcid = NULL, roles = c(), ...) {
   role_names <- credit_roles("name")
 
   # check roles are in list
@@ -68,7 +69,7 @@ author <- function(orcid, surname, given = "", roles = c(), ...) {
     chk_roles <- role_names[credit_roles("abbr") %in% chk_roles]
   }
 
-  orcid <- check_orcid(orcid)
+  if (!is.null(orcid)) orcid <- check_orcid(orcid)
   a <- list(
     orcid = orcid,
     name = list(surname = surname, given = given),